+++ forui/lib/src/theme/style.dart
@@ -11,8 +11,14 @@ import 'package:forui/theme.dart';
 /// These properties are not used directly by Forui widgets. Instead, they are the defaults for the corresponding
 /// properties of widget styles configured via `inherit(...)` constructors.
 final class FStyle with Diagnosticable {
-  /// The form field's style.
-  final FFormFieldStyle formFieldStyle;
+  /// The style for the form field when it is enabled.
+  final FFormFieldStyle enabledFormFieldStyle;
+
+  /// The style for the form field when it is disabled.
+  final FFormFieldStyle disabledFormFieldStyle;
+
+  /// The style for the form field when it has an error.
+  final FFormFieldErrorStyle errorFormFieldStyle;
 
   /// The border radius. Defaults to `BorderRadius.circular(8)`.
   final BorderRadius borderRadius;
@@ -28,7 +34,9 @@ final class FStyle with Diagnosticable {
   /// **Note:**
   /// Unless you are creating a completely new style, modifying [FThemes]' predefined styles should be preferred.
   FStyle({
-    required this.formFieldStyle,
+    required this.enabledFormFieldStyle,
+    required this.disabledFormFieldStyle,
+    required this.errorFormFieldStyle,
     this.borderRadius = const BorderRadius.all(Radius.circular(8)),
     this.borderWidth = 1,
     this.pagePadding = const EdgeInsets.symmetric(vertical: 8, horizontal: 12),
@@ -39,34 +47,37 @@ final class FStyle with Diagnosticable {
     required FColorScheme colorScheme,
     required FTypography typography,
   }) : this(
-          formFieldStyle: FFormFieldStyle.inherit(
-            colorScheme: colorScheme,
-            typography: typography,
-          ),
-        );
+            enabledFormFieldStyle: FFormFieldStyle.inherit(
+              labelColor: colorScheme.primary,
+              descriptionColor: colorScheme.mutedForeground,
+              typography: typography,
+            ),
+            disabledFormFieldStyle: FFormFieldStyle.inherit(
+              labelColor: colorScheme.primary.withOpacity(0.7),
+              descriptionColor: colorScheme.mutedForeground.withOpacity(0.7),
+              typography: typography,
+            ),
+            errorFormFieldStyle: FFormFieldErrorStyle.inherit(
+              labelColor: colorScheme.error,
+              descriptionColor: colorScheme.mutedForeground,
+              errorColor: colorScheme.error,
+              typography: typography,
+            ),);
 
   /// Returns a copy of this [FStyle] with the given properties replaced.
-  ///
-  /// ```dart
-  /// final style = FStyle(
-  ///   borderRadius: BorderRadius.circular(1),
-  ///   borderWidth: 2,
-  /// );
-  ///
-  /// final copy = style.copyWith(borderWidth: 3);
-  ///
-  /// print(copy.borderRadius); // BorderRadius.circular(1)
-  /// print(copy.borderWidth); // 3
-  /// ```
   @useResult
   FStyle copyWith({
-    FFormFieldStyle? formFieldStyle,
+    FFormFieldStyle? enabledFormFieldStyle,
+    FFormFieldStyle? disabledFormFieldStyle,
+    FFormFieldErrorStyle? errorFormFieldStyle,
     BorderRadius? borderRadius,
     double? borderWidth,
     EdgeInsets? pagePadding,
   }) =>
       FStyle(
-        formFieldStyle: formFieldStyle ?? this.formFieldStyle,
+        enabledFormFieldStyle: enabledFormFieldStyle ?? this.enabledFormFieldStyle,
+        disabledFormFieldStyle: disabledFormFieldStyle ?? this.disabledFormFieldStyle,
+        errorFormFieldStyle: errorFormFieldStyle ?? this.errorFormFieldStyle,
         borderRadius: borderRadius ?? this.borderRadius,
         borderWidth: borderWidth ?? this.borderWidth,
         pagePadding: pagePadding ?? this.pagePadding,
@@ -76,7 +87,9 @@ final class FStyle with Diagnosticable {
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {
     super.debugFillProperties(properties);
     properties
-      ..add(DiagnosticsProperty('formFieldStyle', formFieldStyle))
+      ..add(DiagnosticsProperty('enabledFormFieldStyle', enabledFormFieldStyle))
+      ..add(DiagnosticsProperty('disabledFormFieldStyle', disabledFormFieldStyle))
+      ..add(DiagnosticsProperty('errorFormFieldStyle', errorFormFieldStyle))
       ..add(DiagnosticsProperty('borderRadius', borderRadius, defaultValue: BorderRadius.circular(8)))
       ..add(DoubleProperty('borderWidth', borderWidth, defaultValue: 1))
       ..add(DiagnosticsProperty('pagePadding', pagePadding, defaultValue: const EdgeInsets.all(4)));
@@ -87,11 +100,19 @@ final class FStyle with Diagnosticable {
       identical(this, other) ||
       other is FStyle &&
           runtimeType == other.runtimeType &&
-          formFieldStyle == other.formFieldStyle &&
+          enabledFormFieldStyle == other.enabledFormFieldStyle &&
+          disabledFormFieldStyle == other.disabledFormFieldStyle &&
+          errorFormFieldStyle == other.errorFormFieldStyle &&
           borderRadius == other.borderRadius &&
           borderWidth == other.borderWidth &&
           pagePadding == other.pagePadding;
 
   @override
-  int get hashCode => formFieldStyle.hashCode ^ borderRadius.hashCode ^ borderWidth.hashCode ^ pagePadding.hashCode;
+  int get hashCode =>
+      enabledFormFieldStyle.hashCode ^
+      disabledFormFieldStyle.hashCode ^
+      errorFormFieldStyle.hashCode ^
+      borderRadius.hashCode ^
+      borderWidth.hashCode ^
+      pagePadding.hashCode;
 }