+++ forui/lib/src/widgets/text_field/text_field_style.dart
@@ -33,10 +33,10 @@ final class FTextFieldStyle with Diagnosticable {
   final EdgeInsets scrollPadding;
 
   /// The style when this text field is enabled.
-  final FTextFieldNormalStyle enabledStyle;
+  final FTextFieldStateStyle enabledStyle;
 
   /// The style when this text field is enabled.
-  final FTextFieldNormalStyle disabledStyle;
+  final FTextFieldStateStyle disabledStyle;
 
   /// The style when this text field has an error.
   final FTextFieldErrorStyle errorStyle;
@@ -61,21 +61,21 @@ final class FTextFieldStyle with Diagnosticable {
         cursorColor = CupertinoColors.activeBlue,
         contentPadding = const EdgeInsets.symmetric(horizontal: 15, vertical: 15),
         scrollPadding = const EdgeInsets.all(20.0),
-        enabledStyle = FTextFieldNormalStyle.inherit(
+        enabledStyle = FTextFieldStateStyle.inherit(
           contentColor: colorScheme.primary,
           hintColor: colorScheme.mutedForeground,
           focusedBorderColor: colorScheme.primary,
           unfocusedBorderColor: colorScheme.border,
-          formFieldNormaStyle: style.formFieldStyle.enabledStyle,
+          formFieldStyle: style.enabledFormFieldStyle,
           typography: typography,
           style: style,
         ),
-        disabledStyle = FTextFieldNormalStyle.inherit(
+        disabledStyle = FTextFieldStateStyle.inherit(
           contentColor: colorScheme.primary.withOpacity(0.7),
           hintColor: colorScheme.border.withOpacity(0.7),
           focusedBorderColor: colorScheme.border.withOpacity(0.7),
           unfocusedBorderColor: colorScheme.border.withOpacity(0.7),
-          formFieldNormaStyle: style.formFieldStyle.disabledStyle,
+          formFieldStyle: style.disabledFormFieldStyle,
           typography: typography,
           style: style,
         ),
@@ -84,7 +84,7 @@ final class FTextFieldStyle with Diagnosticable {
           hintColor: colorScheme.mutedForeground,
           focusedBorderColor: colorScheme.error,
           unfocusedBorderColor: colorScheme.error,
-          formFieldErrorStyle: style.formFieldStyle.errorStyle,
+          formFieldErrorStyle: style.errorFormFieldStyle,
           typography: typography,
           style: style,
         );
@@ -96,8 +96,8 @@ final class FTextFieldStyle with Diagnosticable {
     Color? cursorColor,
     EdgeInsets? contentPadding,
     EdgeInsets? scrollPadding,
-    FTextFieldNormalStyle? enabledStyle,
-    FTextFieldNormalStyle? disabledStyle,
+    FTextFieldStateStyle? enabledStyle,
+    FTextFieldStateStyle? disabledStyle,
     FTextFieldErrorStyle? errorStyle,
   }) =>
       FTextFieldStyle(